+++ R/download_functions.R
@@ -3,7 +3,7 @@
 #' @format Named list of all download functions
 #' @include lexicon_afinn.R lexicon_loughran.R lexicon_bing.R lexicon_nrc.R
 #' @include dataset_sentence_polarity.R dataset_ag_news.R dataset_dbpedia.R
-#' @include dataset_trec.R dataset_imdb.R
+#' @include dataset_trec.R dataset_imdb.R lexicon_nrc_eil.R
 #' @noRd
 "download_functions"
 download_functions <- list(
@@ -12,6 +12,7 @@ download_functions <- list(
   loughran = download_loughran,
   bing = download_bing,
   nrc = download_nrc,
+  nrc_eil = download_nrc_eil,
   ag_news = download_ag_news,
   dbpedia = download_dbpedia,
   trec = download_trec,