+++ R/process_functions.R
@@ -10,6 +10,7 @@ process_functions <- list(
   loughran = process_loughran,
   bing = process_bing,
   nrc = process_nrc,
+  nrc_eil = process_nrc_eil,
   ag_news = process_ag_news,
   dbpedia = process_dbpedia,
   trec = process_trec,